+++ cirq-core/cirq/contrib/requirements.txt
@@ -9,7 +9,3 @@ opt_einsum
 autoray
 # required for py39 opcodes.
 numba>=0.53.0
-
-# quil import
-pyquil~=2.28.2; python_version < "3.7"
-pyquil~=3.0.0; python_version >= "3.7"
